CV temp/Streamlit App/requirements.txt
streamlit
torch
torchvision
transformers
Pillow
scikit-learn
matplotlib
numpy
tqdm
joblib